requirements.txt
absl-py==1.0.0
appdirs==1.4.4
attrs==21.4.0
axisregistry==0.3.10
babelfont==2.0.2
beautifulsoup4==4.11.1
beziers==0.5.0
booleanOperations==0.9.0
Brotli==1.0.9
browserstack-local==1.2.2
bump2version==1.0.1
bumpfontversion==0.4.1
cattrs==22.1.0
certifi==2022.5.18.1
cffi==1.15.0
cffsubr==0.2.9.post1
charset-normalizer==2.0.12
click==8.1.3
cmarkgfm==0.8.0
collidoscope==0.4.1
colorlog==6.6.0
commandlines==0.4.1
commonmark==0.9.1
compreffor==0.5.1.post1
cu2qu==1.6.7.post1
defcon==0.10.1
dehinter==4.0.0
Deprecated==1.2.13
drawbot-skia==0.4.8
exceptiongroup==1.0.0rc7
font-v==2.1.0
fontbakery==0.8.11
fontdiffenator==0.9.12
fontFeatures==1.7.0
fontmake==3.5.1
fontMath==0.9.3
fontParts==0.10.5
fontPens==0.2.4
fonttools==4.38.0
freetype-py==2.3.0
fs==2.4.16
gfdiffbrowsers==0.1.7
gflanguages==0.4.0
gftools==0.9.21
gitdb==4.0.9
GitPython==3.1.27
glyphsets==0.5.0
glyphsLib==6.2.1
glyphtools==0.8.0
hyperglot==0.3.8
idna==3.3
importlib-resources==5.10.1
Jinja2==3.1.2
lxml==4.8.0
MarkupSafe==2.1.1
munkres==1.1.4
nanoemoji==0.15.1
ninja==1.11.1
numpy==1.22.4
openstep-plist==0.3.0
opentype-sanitizer==8.2.1
opentypespec==1.8.4
packaging==23.1
picosvg==0.20.6
Pillow==9.1.1
pip-api==0.0.29
pngquant-cli==2.17.0.post5
protobuf==3.20.1
psutil==5.9.1
pybind11==2.9.2
pybrowserstack-screenshots==0.1
pycairo==1.21.0
pyclipper==1.3.0.post2
pycparser==2.21
pygit2==1.9.1
PyGithub==1.55
Pygments==2.12.0
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.27.1
resvg-cli==0.22.0.post3
rich==12.4.2
rstr==3.1.0
sh==1.14.2
simplejson==3.17.6
six==1.16.0
skia-pathops==0.7.2
skia-python==87.4
smmap==5.0.0
soupsieve==2.3.2.post1
sre-yield==1.2
statmake==0.5.0
strictyaml==1.6.1
stringbrewer==0.0.1
tabulate==0.8.9
toml==0.10.2
ttfautohint-py==0.5.1
typing_extensions==4.2.0
ufo2ft==2.30.0
ufoLib2==0.13.1
ufolint==1.2.0
uharfbuzz==0.35.0
unicodedata2==14.0.0
Unidecode==1.3.4
urllib3==1.26.9
vharfbuzz==0.2.0
vttLib==0.11.0
wrapt==1.14.1
zipp==3.11.0
zopfli==0.2.2
